pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "async-igdb"
version = "0.0.1"
dependencies = [
  "httpx",
  "pydantic"
]
requires-python = ">=3.12"
authors = [
  {name = "Dax Harris", email = "mvh3875@rit.edu"}
]
description = "Asynchronous wrapper around the IGDB API"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["api", "async", "igdb"]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Programming Language :: Python",
  "License :: OSI Approved :: MIT License",
  "Topic :: Games/Entertainment"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "python-dotenv"
]

[tool.pyright]
reportInvalidTypeForm = false

[project.urls]
Homepage = "https://github.com/iTecAI/async-igdb"
Repository = "https://github.com/iTecAI/async-igdb.git"